moved progressBar out of detailActivity
